Remove Vagrant/cloud compatibility, use bash in setup.sh
Converting a server installation to desktop doesn't seem to work well at
all (performance issues, the installation isn't the same), so abandon
this capability. It is now presumed that this will be run in a complete
desktop installation.

Correct #! line in setup script to use bash explicitly. bashisms are
used but was using POSIX sh. This worked but is bad practice.